refactor(database): make migrations more robust and consistent
- Add conditional checks before dropping/adding columns in users table migration
- Change ethno_id field type from unsignedInteger to unsignedBigInteger for consistency

+++ database/migrations/2026_01_07_073811_make_family_and_siblings_fields_nullable.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->text('office_address')->nullable()->change();
             $table->string('educational_attainment')->nullable()->change();
             $table->integer('income')->nullable()->change();
-            $table->unsignedInteger('ethno_id')->nullable()->change();
+            $table->unsignedBigInteger('ethno_id')->nullable()->change();
         });
 
         Schema::table('fam_siblings', function (Blueprint $table) {
@@ -41,7 +41,7 @@ return new class extends Migration
             $table->text('office_address')->nullable(false)->change();
             $table->string('educational_attainment')->nullable(false)->change();
             $table->integer('income')->nullable(false)->change();
-            $table->unsignedInteger('ethno_id')->nullable(false)->change();
+            $table->unsignedBigInteger('ethno_id')->nullable(false)->change();
         });
 
         Schema::table('fam_siblings', function (Blueprint $table) {
